refactor: reorganize content scanner imports and rename DOMHelper to DOMMessenger
