+++ src/common/services/content-scanner.ts
@@ -1,42 +1,7 @@
 // The 'require.context' feature depends on WebPack (@types/webpack)
-const context: __WebpackModuleApi.RequireContext = require.context('./content-scanners', true, /\.ts$/, 'sync');
-import { CATWikiPageSearchResults, PagesDB } from './database';
+const context: __WebpackModuleApi.RequireContext = require.context('../../content-scanners', true, /\.ts$/, 'sync');
 import DefaultScanner from '@/content-scanners/default-scanner';
-import { IDOMHelperInterface, DOMHelperMessageType } from './domhelper';
-
-export interface IContentScannerPlugin {
-    metaInfo(): string;
-
-    canScanContent(params: IScanParameters): boolean;
-
-    scan(params: IScanParameters): Promise<boolean>;
-}
-
-export interface IScanParameters {
-    domain: string;
-    mainDomain: string;
-    url: string;
-    pagesDb: PagesDB;
-    dom: IDOMHelperInterface;
-    notify: (result: CATWikiPageSearchResults) => void;
-}
-
-// TODO: break this up into per DOMQuery types?
-export interface IContentScanMessage {
-    action: DOMHelperMessageType;
-    id?: string;
-    selector?: string;
-    element?: string;
-    html?: string;
-}
-
-export interface IElementData {
-    tag: string;
-    id: string;
-    className: string;
-    innerText: string;
-    // innerHtml: string;  // can be a bit weighty
-}
+import { IContentScannerPlugin, IScanParameters } from './content-scanner.types';
 
 class ContentScanner {
     private scannerPlugins: IContentScannerPlugin[] = [];
@@ -53,9 +18,7 @@ class ContentScanner {
                 console.log(`Found a plugin that can handle request: ${scannerParameters.domain}`);
                 // TODO: allow multiple handlers ?
                 const didFindPages = await plugin.scan(scannerParameters);
-                if (didFindPages) {
-                    return;
-                }
+                if (didFindPages) return;
             }
         }
         console.log('Using default content scanner');
