+++ src/content-scanners/amazon/amazon-uk.ts
@@ -1,4 +1,4 @@
-import { IContentScannerPlugin, IScanParameters } from '@/content-scanner';
+import { IContentScannerPlugin, IScanParameters } from '@/common/services/content-scanner.types';
 
 class AmazonUKPageScanner implements IContentScannerPlugin {
     metaInfo(): string {
